% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brand_entity}
\alias{brand_entity}
\title{Peaks and Pits brand regex patterns}
\format{
A data frame with brand-related regex patterns.
}
\source{
Internal Peaks and Pits project data
}
\usage{
brand_entity
}
\description{
A data frame of different regex patterns for Peaks and Pits removal of brands.
}
\keyword{datasets}
