% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hf_emotion.R
\name{clean_hf_emotion}
\alias{clean_hf_emotion}
\title{Wrangle distilbert output into a nice format}
\usage{
clean_hf_emotion(df)
}
\arguments{
\item{df}{The dataframe output from the python emotion script}
}
\value{
A dataframe
}
\description{
Wrangle distilbert output into a nice format
}
