% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_emotion_distilbert.R
\name{extract_emotion_distilbert}
\alias{extract_emotion_distilbert}
\title{Wrangle distilbert output into a nice format}
\usage{
extract_emotion_distilbert(df)
}
\arguments{
\item{df}{The dataframe output from the python emotion script}
}
\value{
A dataframe
}
\description{
Wrangle distilbert output into a nice format
}
