% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_emotion_distilroberta.R
\name{extract_emotion_distilroberta}
\alias{extract_emotion_distilroberta}
\title{Wrangle distilroberta output into a nice format}
\usage{
extract_emotion_distilroberta(df)
}
\arguments{
\item{df}{The dataframe output from the python emotion script}
}
\value{
A dataframe
}
\description{
Wrangle distilroberta output into a nice format
}
