% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_emotion_goemotion.R
\name{extract_emotion_goemotion}
\alias{extract_emotion_goemotion}
\title{Wrangle goemotion output into a nice format}
\usage{
extract_emotion_goemotion(df)
}
\arguments{
\item{df}{The dataframe output from the python emotion script}
}
\value{
A dataframe
}
\description{
Wrangle goemotion output into a nice format
}
