% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jh_viz_ngram.R
\name{jh_viz_ngram}
\alias{jh_viz_ngram}
\title{jh_viz_ngram}
\usage{
jh_viz_ngram(tbl_graph, emphasis = TRUE)
}
\arguments{
\item{tbl_graph}{tbl_graph}

\item{emphasis}{emphasis}
}
\value{
ngram plot
}
\description{
jh_viz_ngram
}
