% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_DTMs_parallel.R
\name{make_DTMs_parallel}
\alias{make_DTMs_parallel}
\title{Make Document Term Matrix using parallel processing}
\usage{
make_DTMs_parallel(
  df,
  text_var,
  min_freq = 10,
  hashtags = FALSE,
  mentions = FALSE
)
}
\arguments{
\item{df}{The og dataframe}

\item{text_var}{Column representing text variable}

\item{min_freq}{Minimum frequency}

\item{hashtags}{Keep hashtags?}

\item{mentions}{Keep mentions?}
}
\value{
A list
}
\description{
Make Document Term Matrix using parallel processing
}
