% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner_mask.R
\name{ner_brand_product}
\alias{ner_brand_product}
\title{Mask named entities (brands or products) with a seperate terms for ORG or MISC}
\usage{
ner_brand_product(ner_df, input_df, text_var)
}
\arguments{
\item{ner_df}{The dataframe that has been output by the NER model in huggingface}

\item{input_df}{The dataframe that contains the social post with a document id number}

\item{text_var}{The column containing the text variable}
}
\value{
A dataframe
}
\description{
Mask named entities (brands or products) with a seperate terms for ORG or MISC
}
