% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner_mask.R
\name{ner_mask}
\alias{ner_mask}
\title{Mask named entities (brands or products) with a single term}
\usage{
ner_mask(ner_df, input_df, text_var, mask_term = "mask")
}
\arguments{
\item{ner_df}{The dataframe that has been output by the NER model in huggingface}

\item{input_df}{The dataframe that contains the social post with a document id number}

\item{text_var}{The column containing the text variable}
}
\value{
A dataframe
}
\description{
Mask named entities (brands or products) with a single term
}
