% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_clipboard.R
\name{paste_clipboard}
\alias{paste_clipboard}
\title{paste_clipboard}
\usage{
paste_clipboard(header = FALSE)
}
\arguments{
\item{header}{Logical. Whether the first row should be treated as column headers. Default is FALSE.}
}
\value{
The data from the clipboard get's saved in the R
}
\description{
paste_clipboard
}
\examples{
\dontrun{
df <- paste_clipboard(header = FALSE)
}

}
