% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coherence.R
\name{plot_coherence}
\alias{plot_coherence}
\title{Plot mean and median coherence scores for different values of K topics}
\usage{
plot_coherence(explore_tbl_df, freq_cutoff)
}
\arguments{
\item{explore_tbl_df}{Explore tbl_df from explore_ldas}

\item{freq_cutoff}{Frequency cutoff value}
}
\value{
ggplot line plot
}
\description{
Plot mean and median coherence scores for different values of K topics
}
