% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_pits.R
\name{pp_brands}
\alias{pp_brands}
\title{pp_brands}
\usage{
pp_brands(df, text_var)
}
\arguments{
\item{df}{Input dataframe}

\item{text_var}{Column with posts}
}
\value{
Data Frame or Tibble object with text variable edited inline
}
\description{
pp_brands
}
