% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks_pits.R
\name{pp_products}
\alias{pp_products}
\title{pp_products}
\usage{
pp_products(df, text_var)
}
\arguments{
\item{df}{input dataframe}

\item{text_var}{column with posts}
}
\value{
Data Frame or Tibble object with text variable edited inline
}
\description{
pp_products
}
