% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_entity}
\alias{product_entity}
\title{Peaks and Pits products regex patterns}
\format{
A data frame with product-related regex patterns.
}
\source{
Internal Peaks and Pits project data
}
\usage{
product_entity
}
\description{
A data frame of different regex patterns for Peaks and Pits removal of products.
}
\keyword{datasets}
