% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_capture.R
\name{theme_capture}
\alias{theme_capture}
\title{theme_capture}
\usage{
theme_capture(
  base_size = 11,
  base_family = "GT Walsheim Pro",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  ...
)
}
\arguments{
\item{base_size}{base font size, given in pts}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{...}{}
}
\description{
A ggplot2 based theme for Capture graphics
}
\examples{
p1 <- ggplot(mtcars2) +
geom_point(aes(x = wt, y = mpg, colour = gear)) +
+ theme_capture()

}
