% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brand_entity}
\alias{brand_entity}
\title{Peaks and Pits brand regex patterns}
\format{
\subsection{\code{brand_entity}}{

A data frame of different regex patterns for Peaks and Pits removal of brands
\describe
...
}
}
}
\source{
<
}
\usage{
brand_entity
}
\description{
Peaks and Pits brand regex patterns
}
\keyword{datasets}
