% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_entity}
\alias{product_entity}
\title{Peaks and Pits products regex patterns}
\format{
\subsection{\code{product_entity}}{

A data frame of different regex patterns for Peaks and Pits removal of products
\describe
...
}
}
}
\source{
<
}
\usage{
product_entity
}
\description{
Peaks and Pits products regex patterns
}
\keyword{datasets}
